requirements.txt
streamlit>=1.31.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
hypothesis>=6.0.0
